backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
requests==2.31.0
drf-yasg==1.21.7
python-dotenv==1.0.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
openai==1.3.5
anthropic==0.5.0
tiktoken==0.5.1
channels==4.0.0
channels-redis==4.1.0
django-ratelimit==4.1.0
django-csp==3.7
django-defender==0.0.16
